+++ src/main/index.js
@@ -6,11 +6,9 @@ import fs from 'fs'
 import axios from 'axios'
 import os from 'os'
 import cheerio from 'cheerio'
-import extract from 'extract-zip'
 
-const toml = require('@iarna/toml')
-const { exec } = require('child_process')
-const ws = require('windows-shortcuts')
+import { installVirgil, installDependence, createStartFile, setConfig } from './installation'
+import { readAndParseJSONFile, writeJSONToFile } from './utils.js'
 
 let mainWindow
 function createWindow() {
@@ -172,386 +170,12 @@ function check_if_update(actual, last_version) {
   }
 }
 
-function getVersionVirgil() {
-  return axios.get('https://github.com/ProjecVirgil/VirgilAI/tags').then((response) => {
-    const $ = cheerio.load(response.data)
-    const content = $('body').html()
-    const regex = /v\d+\.\d+\.\d+/
-    const results = content.match(regex)
-    if (results && results.length > 0) {
-      return results[0]
-    } else {
-      throw new Error('Version not found')
-    }
-  })
-}
-
-async function downloadFile(fileUrl, outputLocationPath) {
-  const writer = fs.createWriteStream(outputLocationPath)
-
-  return axios({
-    method: 'get',
-    url: fileUrl,
-    responseType: 'stream'
-  }).then((response) => {
-    response.data.pipe(writer)
-    return new Promise((resolve, reject) => {
-      writer.on('finish', resolve)
-      writer.on('error', reject)
-    })
-  })
-}
-
-function modifyTomlFile(filePath, modifications) {
-  fs.readFile(filePath, 'utf8', (readErr, data) => {
-    if (readErr) {
-      console.error('Error during writing:', readErr)
-      return
-    }
-    try {
-      let config = toml.parse(data)
-      if (!config.tool) {
-        config.tool = {}
-      }
-      if (!config.tool.config_system) {
-        config.tool.config_system = {}
-      }
-      Object.keys(modifications).forEach((key) => {
-        config.tool.config_system[key] = modifications[key]
-      })
-      const modifiedToml = toml.stringify(config)
-      fs.writeFile(filePath, modifiedToml, 'utf8', (writeErr) => {
-        if (writeErr) {
-          console.error('Error during the write phase:', writeErr)
-          return
-        }
-      })
-    } catch (err) {
-      console.error('Error during the analysis or the writing of file', err)
-    }
-  })
-}
-
-function execCommand(command) {
-  return new Promise((resolve, reject) => {
-    exec(command, { shell: true }, (error, stdout) => {
-      if (error) {
-        reject(error)
-      } else {
-        resolve(stdout)
-      }
-    })
-  })
-}
-
 // * ------- SETUP --------
-
-async function installVirgil(event) {
-  try {
-    const lastVersion = await getVersionVirgil()
-    const filename = `${lastVersion}.zip`
-    const pathInstallation = path.join(app.getPath('downloads'), filename)
-    const downloadUrl = `https://github.com/ProjecVirgil/VirgilAI/archive/refs/tags/${lastVersion}.zip`
-
-    await downloadFile(downloadUrl, pathInstallation)
-
-    const data = await readAndParseJSONFile('config.json')
-    await extract(pathInstallation, { dir: data.installation_path })
-
-    event.sender.send('outputcommand', 'COMPLETE')
-  } catch (error) {
-    event.sender.send('outputcommand', 'error ' + error.message)
-  }
-}
-
-async function installDependence(event) {
-  try {
-    const last_version = await getVersionVirgil()
-    const data = await readAndParseJSONFile('config.json')
-
-    const baseDir = path.join(data.installation_path, `VirgilAI-${last_version.replace('v', '')}`)
-
-    // Install Python
-    const pathPython = path.join(baseDir, 'setup')
-    let stdout
-    try {
-      stdout = await execCommand('python -V')
-      if (!(stdout.trim() == 'Python 3.11.7')) {
-        await execCommand(
-          `cd ${pathPython} && python-3.11.7-amd64.exe /quiet InstallAllUsers=1 PrependPath=1`
-        )
-      }
-    } catch (error) {
-      if (error.message.includes('Python was not found')) {
-        // Python not installed, proceed with installation
-        await execCommand(
-          `cd ${pathPython} && python-3.11.7-amd64.exe /quiet InstallAllUsers=1 PrependPath=1`
-        )
-      } else {
-        throw error
-      }
-    }
-    // Setup Python environment
-    const pathPythonEnv = path.join(baseDir)
-    stdout = await execCommand(
-      `cd ${pathPythonEnv} && "C:\\Program Files\\Python311\\python.exe" -m venv virgil-env && .\\virgil-env\\Scripts\\activate.bat && cd setup && pip install -r ./requirements.txt`
-    )
-
-    // Install Poetry dependencies
-    stdout = await execCommand(
-      `cd ${pathPythonEnv} && .\\virgil-env\\Scripts\\activate.bat && poetry install --no-dev`
-    )
-
-    stdout = await execCommand(
-      'winget install ffmpeg -h --accept-package-agreements --accept-source-agreements'
-    )
-
-    event.sender.send('outputcommand', stdout)
-  } catch (error) {
-    event.sender.send('outputcommand', 'error ' + error)
-    console.error(error)
-  }
-}
-
-async function createStartFile(event) {
-  try {
-    const username = os.userInfo().username
-    const last_version = await getVersionVirgil()
-    const data = await readAndParseJSONFile('config.json')
-
-    const path_directory = path.join(
-      data.installation_path,
-      `VirgilAI-${last_version.replace('v', '')}`
-    )
-
-    let batContent = `
-    @echo off
-    cd ${path_directory}
-    call virgil-env\\Scripts\\activate.bat
-    poetry run python launch.py
-    pause
-    `
-
-    const filePath = path.join(path_directory, 'start.bat')
-
-    await fs.promises.writeFile(filePath, batContent)
-
-    ws.create(
-      path.join(
-        'C:',
-        'Users',
-        username,
-        'AppData',
-        'Roaming',
-        'Microsoft',
-        'Windows',
-        'Start Menu',
-        'Programs',
-        'VirgilAI.lnk'
-      ),
-      {
-        target: filePath,
-        desc: 'VirgilAI start file',
-        icon: path.join(
-          data.installation_path,
-          `VirgilAI-${last_version.replace('v', '')}`,
-          'assets',
-          'img',
-          'icon.ico'
-        )
-      }
-    )
-    event.sender.send('outputcommand', 'success')
-  } catch (err) {
-    console.error('Error writing to .bat file:', err)
-    event.sender.send('outputcommand', 'error ' + err.message)
-  }
-}
-
-async function setConfig(event) {
-  try {
-    const username = os.userInfo().username
-    const last_version = await getVersionVirgil()
-    let data = await readAndParseJSONFile('config.json')
-
-    const validateKey = (value) => value.match(/^[a-f0-9]{32}$/i)
-    //PHASE 0
-    if (!validateKey(data.key)) {
-      const key = await createUser()
-      data.key = key
-      await writeJSONToFile('config.json', data)
-    }
-
-    //PHASE 1
-    if (data.startup) {
-      const sourcePath = path.join(
-        'C:',
-        'Users',
-        username,
-        'AppData',
-        'Roaming',
-        'Microsoft',
-        'Windows',
-        'Start Menu',
-        'Programs',
-        'VirgilAI.lnk'
-      )
-      const destinationPath = path.join(
-        'C:',
-        'Users',
-        username,
-        'AppData',
-        'Roaming',
-        'Microsoft',
-        'Windows',
-        'Start Menu',
-        'Programs',
-        'Startup',
-        'VirgilAI.lnk'
-      )
-      fs.copyFileSync(sourcePath, destinationPath)
-    } else {
-      const filePath = path.join(
-        'C:',
-        'Users',
-        username,
-        'AppData',
-        'Roaming',
-        'Microsoft',
-        'Windows',
-        'Start Menu',
-        'Programs',
-        'Startup',
-        'VirgilAI.lnk'
-      )
-      if (fs.existsSync(filePath)) {
-        try {
-          fs.unlinkSync(filePath)
-        } catch (err) {
-          console.error('Error during the remotion of file:', err)
-        }
-      } else {
-        console.log('The file dont exist')
-      }
-    }
-    //PHASE 2
-    modifyTomlFile(
-      path.join(
-        data.installation_path,
-        `VirgilAI-${last_version.replace('v', '')}`,
-        'pyproject.toml'
-      ),
-      { default_start: data.type_interface }
-    )
-    //PHASE 3
-    if (!data.display_console) {
-      const path_directory = path.join(
-        data.installation_path,
-        `VirgilAI-${last_version.replace('v', '')}`
-      )
-      let batContent = `
-      @echo off
-      cd ${path_directory}
-      call virgil-env\\Scripts\\activate.bat
-      poetry run "C:\\Program Files\\python311\\python.exe" launch.pyw
-      `
-      const filePath = path.join(path_directory, 'start.bat')
-
-      await fs.promises.writeFile(filePath, batContent, (err) => {
-        if (err) {
-          console.error('Error writing to .bat file:', err)
-        }
-      })
-
-      await fs.promises.rename(
-        path.join(path_directory, 'launch.py'),
-        path.join(path_directory, 'launch.pyw'),
-        (err) => {
-          if (err) {
-            console.error('Error during the renaming of file:', err)
-          }
-        }
-      )
-    } else {
-      const path_directory = path.join(
-        data.installation_path,
-        `VirgilAI-${last_version.replace('v', '')}`
-      )
-      let batContent = `
-      @echo off
-      cd ${path_directory}
-      call virgil-env\\Scripts\\activate.bat
-      poetry run python launch.py
-      pause
-      `
-      const filePath = path.join(path_directory, 'start.bat')
-
-      await fs.promises.writeFile(filePath, batContent, (err) => {
-        if (err) {
-          console.error('Error writing to .bat file:', err)
-        }
-      })
-      if (fs.existsSync(path.join(path_directory, 'launch.pyw'))) {
-        await fs.promises.rename(
-          path.join(path_directory, 'launch.pyw'),
-          path.join(path_directory, 'launch.py'),
-          (err) => {
-            if (err) {
-              console.error('Error during the renaming of file:', err)
-            }
-          }
-        )
-      }
-    }
-
-    if (data.icon_on_desktop) {
-      const path_directory = path.join(
-        data.installation_path,
-        `VirgilAI-${last_version.replace('v', '')}`
-      )
-      const filePath = path.join(path_directory, 'start.bat')
-
-      ws.create(`C:\\Users\\${username}\\Desktop\\VirgilAI.lnk`, {
-        target: filePath,
-        desc: 'VirgilAI start file',
-        icon: path.join(
-          data.installation_path,
-          `VirgilAI-${last_version.replace('v', '')}`,
-          'assets',
-          'img',
-          'icon.ico'
-        )
-      })
-    } else {
-      fs.access('C:\\Users\\${username}\\Desktop\\VirgilAI.lnk', fs.constants.F_OK, (err) => {
-        const path_directory = path.join(
-          data.installation_path,
-          `VirgilAI-${last_version.replace('v', '')}`
-        )
-        const filePath = path.join(path_directory, 'start.bat')
-        if (err) {
-          console.log('The file dont exist')
-        } else {
-          fs.unlink(filePath, (err) => {
-            if (err) {
-              console.error('Error during the remotion of  file:', err)
-            }
-          })
-        }
-      })
-    }
-
-    event.sender.send('outputcommand', 'success')
-  } catch (error) {
-    console.error(error)
-    event.sender.send('outputcommand', 'error ' + error.message)
-  }
-}
 
 ipcMain.on('runcommand', (event, command) => {
   setTimeout(() => {
     if (command === 'InsVir') {
-      installVirgil(event) //WORK
+      installVirgil(event, app) //WORK
       // event.sender.send('outputcommand', 'success')
     } else if (command === 'InsPy') {
       installDependence(event) //WORK
@@ -578,36 +202,6 @@ ipcMain.on('getJSON', (event, path) => {
     })
 })
 
-function readAndParseJSONFile(filePath) {
-  return new Promise((resolve, reject) => {
-    fs.readFile(filePath, 'utf8', (err, data) => {
-      if (err) {
-        reject(err)
-        return
-      }
-      try {
-        const jsonData = JSON.parse(data)
-        resolve(jsonData)
-      } catch (parseError) {
-        reject(parseError)
-      }
-    })
-  })
-}
-
-function writeJSONToFile(filePath, jsonData) {
-  return new Promise((resolve, reject) => {
-    const dataString = JSON.stringify(jsonData, null, 2)
-    fs.writeFile(filePath, dataString, 'utf8', (err) => {
-      if (err) {
-        reject(err)
-        return
-      }
-      resolve()
-    })
-  })
-}
-
 ipcMain.on('setJSON', (event, data) => {
   writeJSONToFile(data[0], data[1])
     .then(() => {
@@ -619,23 +213,6 @@ ipcMain.on('setJSON', (event, data) => {
 })
 
 //------- CONFIG INIT -------
-
-async function createUser() {
-  const url_base = 'https://virgilapi-production.up.railway.app' + '/api'
-  const url = `${url_base}/createUser`
-  try {
-    const response = await axios.put(url, null, { timeout: 5000 })
-    const userCreated = response.data
-    return userCreated.userId
-  } catch (error) {
-    if (error instanceof axios.RequestError) {
-      console.error("I can't establish a connection, check the network")
-    } else {
-      console.error('User not created')
-    }
-    return 'User not created'
-  }
-}
 
 async function init_config() {
   fs.access('config.json', fs.constants.F_OK, async (err) => {
